requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2

# Redis and Queue
redis==5.0.1
celery==5.3.6
flower==2.0.1

# YouTube download
yt-dlp==2024.1.7
ffmpeg-python==0.2.0
static-ffmpeg==2.5

# Google Cloud
google-cloud-storage==2.14.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# Logging
loguru==0.7.2
